+++ src/test/java/tfw/immutable/ila/shortila/ShortIlaCheck.java
@@ -71,7 +71,7 @@ public final class ShortIlaCheck {
         if (offsetLength < 0) throw new Exception("offsetLength < 0 not allowed");
 
         final StridedShortIla stridedShortIla = StridedShortIlaFromShortIla.create(ila, new short[1000]);
-        final short eps = epsilon < 0.0 ? (short) -epsilon : (short) epsilon;
+        final short eps = epsilon < 0.0 ? (short) -epsilon : epsilon;
         final short neps = (short) -eps;
         final Random random = new Random(0);
         final int ilaLength = ila.length() + offsetLength <= Integer.MAX_VALUE
@@ -113,7 +113,7 @@ public final class ShortIlaCheck {
 
         final StridedShortIla stridedTarget = StridedShortIlaFromShortIla.create(target, new short[1000]);
         final StridedShortIla stridedActual = StridedShortIlaFromShortIla.create(target, new short[1000]);
-        final short eps = epsilon < 0.0 ? (short) -epsilon : (short) epsilon;
+        final short eps = epsilon < 0.0 ? (short) -epsilon : epsilon;
         final short neps = (short) -eps;
         final Random random = new Random(0);
         final int ilaLength = stridedTarget.length() + addlOffsetLength <= Integer.MAX_VALUE
